src/textforensics/pipelines/training.py
"""Training Pipeline for TextForensics"""

import torch
import torch.nn as nn
from torch.utils.data import DataLoader
from typing import Dict, Any, Optional
import logging
from pathlib import Path
import wandb
from omegaconf import DictConfig
import hydra

from ..models import TextForensicsUnifiedModel
from ..utils import setup_logging, save_checkpoint, load_checkpoint


class TrainingPipeline:
    """Training pipeline for TextForensics models"""
    
    def __init__(self, config: DictConfig):
        self.config = config
        self.logger = setup_logging(config)
        
        # Initialize model
        self.model = self._build_model()
        
        # Initialize data loaders (placeholder)
        self.train_loader, self.val_loader = self._build_dataloaders()
        
        # Initialize optimizer and scheduler
        self.optimizer = self._build_optimizer()
        self.scheduler = self._build_scheduler()
        
        # Initialize loss functions
        self.loss_functions = self._build_loss_functions()
        
        # Setup device
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model.to(self.device)
        
        # Initialize tracking
        self.current_epoch = 0
        self.best_val_loss = float('inf')
        self.patience_counter = 0
        
    def _build_model(self) -> nn.Module:
        """Build model from config"""
        # Placeholder - would use hydra.utils.instantiate in real implementation
        from ..models.base import TextForensicsConfig
        from ..models.unified_model import TextForensicsUnifiedModel
        
        config = TextForensicsConfig()
        return TextForensicsUnifiedModel(config)
        
    def _build_dataloaders(self) -> tuple:
        """Build data loaders from config"""
        # Placeholder implementation
        return None, None
        
    def _build_optimizer(self):
        """Build optimizer from config"""
        return torch.optim.AdamW(self.model.parameters(), lr=5e-5)
        
    def _build_scheduler(self):
        """Build learning rate scheduler from config"""
        return torch.optim.lr_scheduler.CosineAnnealingLR(self.optimizer, T_max=10)
        
    def _build_loss_functions(self) -> Dict[str, nn.Module]:
        """Build loss functions for each task"""
        return {"classification": nn.CrossEntropyLoss()}
        
    def run(self) -> Dict[str, Any]:
        """Run the complete training pipeline"""
        
        self.logger.info("Starting training pipeline...")
        self.logger.info(f"Training on device: {self.device}")
        
        # Placeholder training loop
        for epoch in range(self.config.pipeline.max_epochs):
            self.current_epoch = epoch
            
            # Training phase (placeholder)
            train_loss = 0.5  # Dummy loss
            
            # Validation phase (placeholder)
            val_loss = 0.4  # Dummy loss
            
            # Early stopping check
            if val_loss < self.best_val_loss:
                self.best_val_loss = val_loss
                self.patience_counter = 0
            else:
                self.patience_counter += 1
                
            if self.patience_counter >= self.config.pipeline.patience:
                break
                
        return {
            'best_val_loss': self.best_val_loss,
            'final_epoch': self.current_epoch
        }
