requirements.txt
# Core dependencies
torch>=1.13.0
transformers>=4.25.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytorch-lightning>=1.8.0

# ML and Data
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
datasets>=2.8.0
tokenizers>=0.13.0

# Visualization and UI
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
rich>=12.0.0

# CLI and API
typer>=0.7.0
fastapi>=0.85.0
uvicorn>=0.18.0

# NLP Libraries
nltk>=3.7
spacy>=3.4.0

# Utilities
tqdm>=4.64.0
wandb>=0.13.0
python-dotenv>=0.20.0
